Refactor touch so less mix up
